requirements.txt
boto3==1.16.55
botocore==1.19.55
cryptography==3.3.1
openstacksdk==0.52.0
PyYAML==5.3.1
requests==2.25.1
